+++ bank/src/main/java/com/fisa/bank/common/config/security/resource/LoginSuccessHandler.java
@@ -1,4 +1,4 @@
-package com.fisa.bank.common.config.security.authorization;
+package com.fisa.bank.common.config.security.resource;
 
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
