+++ bank/src/main/java/com/fisa/bank/common/presentation/util/SpringRequesterInfo.java
@@ -9,7 +9,7 @@ import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Component;
 
 import com.fisa.bank.common.application.util.RequesterInfo;
-import com.fisa.bank.common.config.security.authorization.JwtAuthentication;
+import com.fisa.bank.common.config.security.resource.JwtAuthentication;
 import com.fisa.bank.user.persistence.entity.id.UserId;
 
 @Component
